Fixed stupidity with method signature...
git-svn-id: https://subversion.assembla.com/svn/d2bs@440 080cfd25-d5d9-4c59-82ea-70d02a200d8b

+++ trunk/D2Helpers.h
@@ -85,6 +85,6 @@ DWORD __fastcall D2CLIENT_GetMinionCount(UnitAny* pUnit, DWORD dwType);
 VOID __fastcall D2CLIENT_HostilePartyUnit(RosterUnit* pUnit, DWORD dwButton);
 VOID __fastcall D2CLIENT_TakeWaypoint(DWORD dwWaypointId, DWORD dwArea);
 
-DWORD GetDistance(DWORD x1, DWORD y1, DWORD x2, DWORD y2, DistanceType type = Euclidean);
+double GetDistance(DWORD x1, DWORD y1, DWORD x2, DWORD y2, DistanceType type = Euclidean);
 
 #endif
